Specify the client protocol, and test that the specification is true
M10 task 1. KimmyDB has had a protocol since M1 and nothing had written it
down: no client could be generated, and nothing failed when a route drifted.

`docs/openapi.yaml` is that document — OpenAPI 3.1 over every route the
router registers, with the Extended JSON v2 conventions, the auth flow, the
error envelope, cursor semantics and the WebSocket frame shapes. `/mcp` is
deliberately absent: a different protocol, config-gated, specified by MCP.

The document is not the deliverable. `crates/kimmy-api/tests/openapi.rs` is:

  - inventory, both directions — every registered route is described and
    every described operation registered;
  - behaviour — every documented operation driven against a real server over
    a real socket, each response validated against its declared schema;
  - coverage — the run ends by asserting every documented operation was
    actually exercised, so a specification entry nothing executes cannot be
    added.

Hand-written rather than generated (ADR-056). Thirty-eight of thirty-nine
handlers return `Json<Value>`, so `utoipa` would automate the half a text
scan already checks and leave every response schema hand-written anyway —
and a generated document cannot disagree with the code, which is exactly the
disagreement worth having.

Three things it found on its first runs:

  - `PUT /docs/{id}` returned `matched` and `modified` as booleans while
    `/update` and `/find_and_modify` return counts — one field name, two
    types, on one protocol, because the route serialized `WriteOutcome`'s
    bools straight to the wire. The route had no integration test at all.
    Normalized to counts by decision, before any client exists to break;
    `upserted` stays a boolean because it is one. It has a test now.
  - `GET /v1/users` returns names, not user objects.
  - The M8 inventory check matched `.route("` at the start of a line, so it
    skipped the three registrations rustfmt breaks across lines, including
    `/docs/{id}`. Green for two milestones while never checking the busiest
    route on the API. There is one scanner now, and it covers the prose
    reference too.

Both new dependencies are dev-only; `check-native-deps.sh` still reports
`cc` alone. Verified beyond the suite against a release node, with a Python
drive that parses the specification itself and checked 32 real responses.

+++ crates/kimmy-api/src/exec.rs
@@ -470,9 +470,17 @@ pub fn replace(
     let doc_id = parse_id(id)?;
     let doc = json_to_document(document)?;
     let outcome = state.engine.replace(&meta, &doc_id, doc, upsert)?;
+    // Counts, not booleans, even though a replace touches at most one document.
+    //
+    // `WriteOutcome` is three bools and this route used to serialize them
+    // straight through, so `matched` was `true` here and `3` on `/update` —
+    // the same field name carrying two types on one protocol. Nothing stated
+    // the type, so nothing disagreed until the specification's contract test
+    // drove both routes and compared them (ADR-056). `upserted` stays a
+    // boolean because it genuinely is one.
     Ok(json!({
-        "matched": outcome.matched,
-        "modified": outcome.modified,
+        "matched": u8::from(outcome.matched),
+        "modified": u8::from(outcome.modified),
         "upserted": outcome.upserted,
     }))
 }
